requirements.txt
aiofiles==0.6.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
config==0.5.0.post0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-import-export==2.5.0
et-xmlfile==1.0.1
gunicorn==20.1.0
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.10
MarkupPy==1.14
multidict==5.0.0
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.3
patsy==0.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
rfc3986==1.4.0
sanic==19.12.5
Sanic-Cors==0.10.0.post3
Sanic-Plugins-Framework==0.9.5
scipy==1.6.2
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
statsmodels==0.12.2
tablib==3.0.0
websockets==8.1
xlrd==2.0.1
xlwt==1.3.0
